docs(t13-f01): record the build pass status, verification runs, and implementation deviations

+++ crates/v3-cli/src/bench.rs
@@ -3853,7 +3853,7 @@ mod tests {
                 .expect("an opportunity block");
             assert_eq!(
                 opportunities.births,
-                u64::from(reading.depth) * u64::from(drift.lineages)
+                reading.depth * u64::from(drift.lineages)
             );
             assert_eq!(
                 opportunities.attempted,
